[GLUTEN-9522][VL] Refactor date & math scalar functions into new suites
